# projects/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "projects-service dev",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/object_status/{id}/towers": {
      "get": {
        "tags": [
          "ObjectStatus"
        ],
        "summary": "Get Object Status With Towers List By Id",
        "description": "Gets object status with towers list by id",
        "operationId": "get_object_status_with_towers_list_by_id_api_object_status__id__towers_get",
        "parameters": [
          {
            "description": "The UUID id of the object status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the object status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IObjectStatusWithTowersListRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tower/list": {
      "get": {
        "tags": [
          "Towers"
        ],
        "summary": "Get Towers List",
        "description": "Gets a paginated list of towers",
        "operationId": "get_towers_list_api_tower_list_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1.0,
              "type": "integer",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100.0,
              "minimum": 1.0,
              "type": "integer",
              "default": 50
            },
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponsePaginated_ITowerRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tower/{id}": {
      "get": {
        "tags": [
          "Towers"
        ],
        "summary": "Read Tower By Id",
        "description": "Gets a tower by id",
        "operationId": "read_tower_by_id_api_tower__id__get",
        "parameters": [
          {
            "description": "The UUID id of the tower",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the tower",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_ITowerRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Towers"
        ],
        "summary": "Update Tower",
        "description": "Updates a tower by its id",
        "operationId": "update_tower_api_tower__id__put",
        "parameters": [
          {
            "description": "The UUID id of the tower",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the tower",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ITowerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPutResponseBase_ITowerRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Towers"
        ],
        "summary": "Delete Tower",
        "description": "Deletes a tower by its id",
        "operationId": "delete_tower_api_tower__id__delete",
        "parameters": [
          {
            "description": "The UUID id of the tower",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the tower",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IDeleteResponseBase_ITowerRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tower": {
      "post": {
        "tags": [
          "Towers"
        ],
        "summary": "Create Tower",
        "description": "Creates a new tower",
        "operationId": "create_tower_api_tower_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ITowerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPostResponseBase_ITowerRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "IDeleteResponseBase_ITowerRead_": {
        "title": "IDeleteResponseBase[ITowerRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data deleted correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/ITowerRead"
          }
        }
      },
...
      "IGetResponseBase_IObjectStatusWithTowersListRead_": {
        "title": "IGetResponseBase[IObjectStatusWithTowersListRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/IObjectStatusWithTowersListRead"
          }
        }
      },
...
      "IGetResponseBase_ITowerRead_": {
        "title": "IGetResponseBase[ITowerRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/ITowerRead"
          }
        }
      },
...
      "IGetResponsePaginated_ITowerRead_": {
        "title": "IGetResponsePaginated[ITowerRead]",
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/PageBase_DataType__ITowerRead_"
          }
        }
      },
...
      "IObjectStatusWithTowersListRead": {
        "title": "IObjectStatusWithTowersListRead",
        "required": [
          "name",
          "id"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string",
            "format": "uuid"
          },
          "towers_entities": {
            "title": "Towers Entities",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ITowerWithoutObjectStatusId"
            },
            "default": []
          }
        }
      },
...
      "IPostResponseBase_ITowerRead_": {
        "title": "IPostResponseBase[ITowerRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data created correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/ITowerRead"
          }
        }
      },
...
      "IPutResponseBase_ITowerRead_": {
        "title": "IPutResponseBase[ITowerRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data updated correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/ITowerRead"
          }
        }
      },
...
      "ITowerCreate": {
        "title": "ITowerCreate",
        "required": [
          "name",
          "subproject_id"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "subproject_id": {
            "title": "Subproject Id",
            "type": "string",
            "format": "uuid"
          },
          "status_id": {
            "title": "Status Id",
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ITowerRead": {
        "title": "ITowerRead",
        "required": [
          "name",
          "subproject_id",
          "id"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "subproject_id": {
            "title": "Subproject Id",
            "type": "string",
            "format": "uuid"
          },
          "status_id": {
            "title": "Status Id",
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "title": "Id",
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ITowerUpdate": {
        "title": "ITowerUpdate",
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "subproject_id": {
            "title": "Subproject Id",
            "type": "string",
            "format": "uuid"
          },
          "status_id": {
            "title": "Status Id",
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ITowerWithoutObjectStatusId": {
        "title": "ITowerWithoutObjectStatusId",
        "required": [
          "id",
          "name",
          "subproject_id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "subproject_id": {
            "title": "Subproject Id",
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "PageBase_DataType__ITowerRead_": {
        "title": "PageBase[DataType][ITowerRead]",
        "required": [
          "items",
          "total",
          "pages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ITowerRead"
            }
          },
          "total": {
            "title": "Total",
            "minimum": 0.0,
            "type": "integer"
          },
          "page": {
            "title": "Page",
            "minimum": 1.0,
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "minimum": 1.0,
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "next_page": {
            "title": "Next Page",
            "type": "integer"
          },
          "previous_page": {
            "title": "Previous Page",
            "type": "integer"
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
...
    }
  }
}